requirements.txt
einops==0.8.2
fastapi==0.129.0
fvcore==0.1.5.post20221221
gradio==6.5.1
logger==1.4
matplotlib==3.10.8
numpy==2.4.2
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
pandas==3.0.0
Pillow==12.1.1
pydicom==3.0.1
Requests==2.32.5
timm==0.9.0
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
uvicorn==0.40.0
